+++ src/vbnigmm/backend/current.py
@@ -1,2 +1,1 @@
-from .constant import *
 from .tensorflow import *
