+++ src/vbnigmm/backend/search.py
