+++ src/vbnigmm/backend/tensorflow/__init__.py
@@ -1,4 +1,3 @@
 from .ops import *
 from .ufunc import *
 from .keras import *
-from .gamma import *
