+++ src/vbnigmm/distributions/base.py
@@ -1,4 +1,4 @@
-from ..backend import current as tk
+from .. import backend as tk
 
 
 class Dist(object):
