+++ src/vbnigmm/linalg/scalar.py
@@ -1,4 +1,4 @@
-from ..backend import current as tk
+from .. import backend as tk
 from .base import Base
