+++ src/vbnigmm/linalg/vector.py
@@ -1,4 +1,4 @@
-from ..backend import current as tk
+from .. import backend as tk
 
 from .base import Base
 from .matrix import InfMatrix, mul_matrix
