+++ src/vbnigmm/mixture/dpm.py
@@ -1,4 +1,4 @@
-from ..backend import current as tk
+from .. import backend as tk
 
 from ..distributions.base import Dist
 from ..distributions.dirichlet import Dirichlet
