+++ src/vbnigmm/mixture/utils.py
@@ -1,4 +1,5 @@
-from ..backend import current as tk
+from .. import backend as tk
+
 from ..distributions.base import Dist
